+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added `--random-extra-roles` option to `start`. ([#30](https://github.com/hjwylde/werewolf/issues/30))
 * Restructured library modules to only export relevant functions. ([#11](https://github.com/hjwylde/werewolf/issues/11))
 * Added the Bear Tamer role. ([#45](https://github.com/hjwylde/werewolf/issues/45))
+* Added a `circle` command. ([#45](https://github.com/hjwylde/werewolf/issues/45))
 
 *Revisions*